Parser.py: use argparse
Use argparse for command line arguments. This gives us the online help and
will give use more flexibility to add command line options.

We preserve the current behaviour with up to two positional arguments for
the ekl and seq input files, and up to two arguments for the key:value
search.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>