Parser.py: handle empty group
Currently the parser tries to record the test service as 'group' and the
test protocol as 'test set'.

However, some tests have only a protocol and no service (e.g.
LoadedImageProtocolTest). In this case, the parser will record the protocol
as 'group' and the full line dictionary as 'test set'.

Fix this case by recording an empty 'group' and the protocol as 'test set',
more in line with the generic case.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>