Add configuration file
Add a --config option to Parser.py, to use a configuration file.

This configuration file can contain a list of transformations to apply to
the test results. This allows to mark tests as false positives or to waive
failures.

We describe the YAML configuration file format in details in the README.md,
and we also add a minimal EBBR.yaml configuration file.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>