cat.sh: use iconv
Currently we `cat' all the discrete .ekl log files into a big, single file
directly as binaries. This adds UTF-16 two bytes BOM everywhere in the
middle of the resulting cat_summary.ekl.

Fix that with iconv: we convert all the log files to UTF-8 on the fly and
back to UTF-16 while concatenating.

While at it, handle an optional command line argument to specify the
folder to search for logs.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>