Better handling of skipped tests sets
A skip test set is a test set, which was considered but had zero of its
tests run. This is not the same as a dropped test set, which we do not find
in the log at all.
Generate artificial test entries with the result set to SKIPPED to keep
track of the former.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>