EBBR.yaml: add configuration with ignored and waived
We add a configuration file with rules for testing the EBBR specification.

Some tests correspond to cases not mandated by the EBBR specification and
we want to ignore them. Some other tests are either already fixed with a
recent U-Boot or still present with the sandbox or qemu, and we want to
issue waivers for those.

                  Original result          Forced result
  False-positive  FAILURE/WARNING/DROPPED  IGNORED
          Waived  FAILURE/WARNING/SKIPPED  WAIVED

The list has been established in part by comparing SCT runs done with the
EBBR.seq sequence file, running on arm64 U-Boot sandbox and qemu, from
commit ddaa94978583 and version v2021.04-rc2.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>