EBBR.yaml: ignore skipped simple text input ex test
Simple Text Input Ex Protocol tests can be skipped when the SCT is unable
to locate the console device bound to the protocol. This is the case with
U-Boot.

EBBR has requirements on the console, but it does not specifically require
that a device be always bound to the Simple Text Input Ex Protocol.

Simply ignore this skipped test instead of waiving it.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>