schema: constrain a bit more
We know all the valid values for many fields of the configuration rules.
Refine the schema to constrain the EBBR.yaml config file a bit more.

This means we are giving up the ability to mention only a sub-string for
those fields in the config file, which we were not doing anyway.

Also, forbid additional properties where we can.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>