Add an SIE configuration
Add a configuration to process SCT results from the Security Interface
Extension ACS.

The rules in this new configuration can override tests results with a new
`KNOWN RPMB LIMITATION' result.

Add the necessary enums to our schema and adjust the sequence files
database to use the new configuration automatically when the SIE ACS is
detected.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>