parser: add meta-data
Collect a few meta-data during execution, such as the date, the parser
command line, the sequence file identification and the git commit of the
parser when possible.

We add a new `--print-meta' command line option to display the collected
meta-data.

We output the collected meta-data to the generated files for the following
formats: md, yaml and template.

Also, we gracefully ignore the meta-data when reading back a result.md file
with meta-data.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>