Re-classify size limitations
The SIE errors, which we originally classified as "known rpmb limitations"
can happen not only with an eMMC but in the more general case with a small
storage for UEFI variables. This is typically the case with OP-TEE.
Let's re-classify those errors as "known size limitations". This is just
renaming at this stage.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>