parser.py: handle some missing keys
In some cases, the SCT can generate a special test entry to notify of
"abnormal hang or stop". This breaks the parsing and results in the
following exception:

  Traceback (most recent call last):
    File "parser.py", line 1085, in <module>
      cross_check = read_log_and_seq(args.log_file, args.seq_file)
    File "parser.py", line 824, in read_log_and_seq
      return combine_dbs(db1, db2)
    File "parser.py", line 695, in combine_dbs
      if x['set guid'] not in s:
         ~^^^^^^^^^^^^
  KeyError: 'set guid'

Detect and create some missing keys to repair parsing.

In this specific case, the failed test ends up being classified as
spurious.

While at it, add a unit test.

Reported-by: Ann Cheng <ann.cheng@arm.com>
Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>